+++ src/controllers/utils/error.messages.ts
@@ -1,6 +1,8 @@
 enum ErrorMessagesForUsers {
-  studentNotFound = 'The requested student does not exist.',
-  teacherNotFound = 'The requested teacher does not exist.'
+  errorUpdating = 'Hubo un error generando la contraseña del usuario.',
+  studentNotFound = 'El estudiante no se encuentra registrado en el padrón.',
+  teacherNotFound = 'El docente no se encuentra registrado en el padrón.',
+  userHasNotMail = 'Ud. no tiene un correo registrado en el padrón electoral.\n Se ha notificado al CEUNI, se comunicarán con Ud. lo antes posible. Disculpe las molestias.'
 }
 
 export { ErrorMessagesForUsers as EFU }