+++ src/dto-interfaces/user.dto.ts
@@ -1,4 +1,6 @@
 export interface DtoUser {
   documentNumber?: string
   documentType?  : string
+  id?            : string
+  password?      : string
 }