+++ changelog.md
@@ -1,5 +1,14 @@
 # Rhoades Backend
 
+## Version 0.2.0
+
+- Implemented:
+  - Password generator.
+  - Encryption of the password.
+  - Endpoint to generate the password.
+  - Notification to the user about the password.
+  - `id` to the `DtoUser` and `IUser` interfaces.
+
 ## Version 0.1.0
 
 - Project Initialization.